Cosas varias
Queda pendiente revisar la generacion de la factura al cliente al final de la funcionalidad ingresar vehiculo

+++ src/gestorAplicacion/personas/Cliente.java
@@ -15,7 +15,7 @@ public class Cliente extends Persona implements Serializable {
     private ArrayList<Vehiculo> vehiculos = new ArrayList<Vehiculo>();
     private Factura factura;
     
-	public Cliente(String nombre, long cedula, int telefono, String correo, String direccion, boolean discapacitado) {
+	public Cliente(String nombre, long cedula, long telefono, String correo, String direccion, boolean discapacitado) {
 		super(nombre, cedula, telefono, correo, direccion);
 		this.discapacitado = discapacitado;
 	}
